feat(autoware_tensorrt_vad): move VadConfig to vad_config.hpp
