+++ AV-Solutions/vad-trt/app/autoware_tensorrt_vad/include/autoware/tensorrt_vad/net.hpp
@@ -29,33 +29,11 @@
 
 #include "tensor.hpp"
 #include "ros_vad_logger.hpp"
+#include "vad_config.hpp"
 
 #include <autoware/tensorrt_common/tensorrt_common.hpp>
 
 namespace autoware::tensorrt_vad {
-
-// NetworkIO configuration parameters
-struct VadConfig
-{
-  int32_t num_cameras;
-  int32_t bev_h, bev_w;
-  int32_t bev_feature_dim;
-  int32_t num_decoder_layers;
-  int32_t prediction_num_queries;
-  int32_t prediction_num_classes;
-  int32_t prediction_bbox_pred_dim;
-  int32_t prediction_trajectory_modes;
-  int32_t prediction_timesteps;
-  int32_t planning_ego_commands;
-  int32_t planning_timesteps;
-  int32_t can_bus_dim;
-  int32_t target_image_width;
-  int32_t target_image_height;
-  int32_t downsample_factor;
-  int32_t map_num_queries;
-  int32_t map_num_class;
-  int32_t map_points_per_polylines;
-};
 
 // Standalone functions for NetworkIO generation and engine building
